% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dens_means}
\alias{dens_means}
\title{Seasonal mean log-densities}
\format{
A tibble with 4 rows (seasons) and 4 columns:
\describe{
\item{season}{Factor. Season (winter, spring, summer, fall).}
\item{log.bm.imp.mean}{Numeric. Mean log-density for blue whale.}
\item{log.bp.imp.mean}{Numeric. Mean log-density for fin whale.}
\item{log.mn.imp.mean}{Numeric. Mean log-density for humpback whale.}
}
}
\usage{
dens_means
}
\description{
Seasonal mean log-density for each whale species, used to back-transform
log-ratio fitted values to the original density scale.
}
\keyword{datasets}
