% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_asvs}
\alias{example_asvs}
\title{Example ASVs}
\format{
A character vector of length 500.
}
\usage{
example_asvs
}
\description{
Character vector of the 500 ASV identifiers included in \code{example_16s}.
}
\keyword{datasets}
